Refactor useWindowOnNearBottom hook
