+++ src/hooks/index.ts
@@ -1,3 +1,3 @@
 export { default as useOnClickOutside } from './click-outside'
 export { default as useOnKeypress } from './key-press'
-export { default as useWindowOnNearBottom } from './use-window-near-bottom'
+export { default as useWindowOnNearBottom } from './window-near-bottom'
